create missing magic images on editor load
If there are missing fileadmin/_processed_ images, the will not be created on editor load and therefore will be displayed as empty images inside the editor.

We need to implement a hook / signal that processes all images before the editor is visible to users. This is basically the same process as if you hit save. After saving, the RteImagesDbHook gets called and all missing processed files will be created.